Added new configuration to limit requests by queue *and* stream.
Prior to this commit, there were request limits for users, subs and
mods, and then a flag `LimitUserRequestsToSession` which toggled whether
those limits applied to the queue or to the session.

This commit implements an extra set of numerical limits so that requests
can be limited by both queue and total session simultaneously. For
example, you can set it up so that a user can only request 10 songs
total in a given session, and simultaneously they cannot have more than
2 requests in queue at a time.

The feature is implemented in a backwards compatible way, such that by
default there is no total-session limit.

Testing done:
* Ran BeatSaber and confirmed user could not exceed the queue-limit nor
  the total-session limit.