Fixed bug where the number of requests in queue could become negative.
To implement requests limits for songs in queue, the program uses a
counter that increments every time a user submits a request and
decrements every time their song is played.

Let's say that the broadcaster has set a limit of up to 2 songs in
queue. There's an edge case where a user can submit 2 songs (and thus
have their counter be set to 2), and then the broadcaster quits the game
(or it crashes), and then reloads the game. The songs in the queue will
be preserved, but the counters get reset to zero. Then, if the
broadcaster plays the user's two songs that are still in queue, the
counter gets decremented twice and will be set to -2. At this point, the
user could make 4 song requests, bypassing the 2 limit that the
broadcaster intended.